+++ database/migrations/2021_11_04_150700_create_product_enter_detail_table.php
@@ -20,7 +20,6 @@ class CreateProductEnterDetailTable extends Migration
             $table->integer('quantity');
             $table->string('notes')->nullable();
             $table->timestamps();
-            $table->softDeletes();
         });
     }